[nagendra|rahit] mocked writeFile,renamed function
What changes you have made?
mocked writeFile in test
renamed isCorrectPath to doesFileNotPresent
checked for file present on loading
removed unnecessary variable in serveAllTodo

Why this change is necessary?
to not allow tests to write to file
to meant what it does

What is the impact of the change?
test not deal with file writting
for readability